feat: TeamSp 변수 분리 (기본정보, 호스트 지정템플릿)
